Moving EntityLineage Utils to separate utils file.
